requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
line-bot-sdk>=3.0.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
